Refactored code, to split the Injector executable class from the class that holds information about what should be injected.
